requirements.txt

openai
transformers
torch
Pillow
opencv-python
pytesseract
moviepy
pandas
requests
